all: RoundRobin doesn't need to be a pointer
It's a struct hoding three integers, which means the space of three
pointers. And we don't need it to be a pointer for it to work properly.

While at it, merge round_robin* files with api_definition*. The former
were tiny and only used as part of APISpec.

Updates #334.
