+++ api_definition.go
@@ -124,7 +124,7 @@ type APISpec struct {
 	Health                   HealthChecker
 	JSVM                     *JSVM
 	ResponseChain            []TykResponseHandler
-	RoundRobin               *RoundRobin
+	RoundRobin               RoundRobin
 	URLRewriteEnabled        bool
 	CircuitBreakerEnabled    bool
 	EnforcedTimeoutEnabled   bool
@@ -1116,4 +1116,32 @@ func (a *APISpec) GetVersionData(r *http.Request) (*apidef.VersionInfo, []URLSpe
 
 	return &version, versionRxPaths, versionWLStatus, StatusOk
 
+}
+
+type RoundRobin struct {
+	pos, max, cur int
+}
+
+func (r *RoundRobin) SetMax(max int) {
+	if r.max = max; r.max < 0 {
+		r.max = 0
+	}
+
+	// Can't have a new list substituted that's shorter
+	if r.cur > r.max {
+		r.cur = 0
+	}
+	if r.pos > r.max {
+		r.pos = 0
+	}
+}
+
+func (r *RoundRobin) SetLen(len int) { r.SetMax(len - 1) }
+
+func (r *RoundRobin) GetPos() int {
+	r.cur = r.pos
+	if r.pos++; r.pos > r.max {
+		r.pos = 0
+	}
+	return r.cur
 }
