+++ api_definition_test.go
@@ -515,3 +515,14 @@ func TestGetAPISpecsDashboardSuccess(t *testing.T) {
 		t.Error("Should return array with one spec", ApiSpecRegister)
 	}
 }
+
+func TestRoundRobin(t *testing.T) {
+	rr := RoundRobin{}
+	rr.SetMax(2)
+
+	for _, want := range []int{0, 1, 2, 0} {
+		if got := rr.GetPos(); got != want {
+			t.Errorf("RR Pos wrong: want %d got %d", want, got)
+		}
+	}
+}
